Cosas de enemigos y eso

+++ Logica/src/main/java/com/example/logica/Vector2D.java
@@ -2,10 +2,10 @@ package com.example.logica;
 
 public class Vector2D {
 
-    public Vector2D(int x, int y) {
+    public Vector2D(float x, float y) {
         _x = x;
         _y = y;
     }
 
-    int _x, _y;
+    float _x, _y;
 }
